+++ main.cpp
@@ -79,7 +79,7 @@ int main(int argc, char *argv[])
         break;
     }
 
-    int32_t level = static_cast<int32_t>(atof(argv[2]) * AudioTrack::UNIT_LEVEL);
+    uint16_t level = static_cast<uint16_t>(atof(argv[2]) * AudioTrack::UNIT_LEVEL);
 
     if (!track.start(wav_file, mode, level)) {
         fprintf(stderr, "Cannot play file\n");
